+++ test/integration/spec/rtcpeerconnection.js
@@ -45,7 +45,7 @@ const sdpSemanticsValues = isFirefox
   ? [null]  // Unified Plan
   : sdpSemanticsIsSupported
     ? ['plan-b', 'unified-plan']
-    : ['plan-b'];
+    : ['currentDirection' in RTCRtpTransceiver.prototype ? 'unified-plan' : 'plan-b'];
 
 sdpSemanticsValues.forEach(sdpSemantics => {
 
@@ -164,7 +164,7 @@ describe(description, function() {
     });
   });
 
-  (isSafari ? describe.skip : describe)('#createOffer, called twice from signaling state "stable" without calling #setLocalDescription', () => {
+  (isSafari && sdpSemantics === 'plan-b' ? describe.skip : describe)('#createOffer, called twice from signaling state "stable" without calling #setLocalDescription', () => {
     let offer1;
     let offer2;
 
@@ -193,7 +193,7 @@ describe(description, function() {
     });
   });
 
-  (isSafari ? describe.skip : describe)('#createAnswer, called twice from signaling state "stable" without calling #setLocalDescription', () => {
+  (isSafari && sdpSemantics === 'plan-b' ? describe.skip : describe)('#createAnswer, called twice from signaling state "stable" without calling #setLocalDescription', () => {
     let answer1;
     let answer2;
 
@@ -259,8 +259,11 @@ describe(description, function() {
 
       // Here is a dumb way to change SSRCs: just delete any SSRC groups, then
       // strip the leading digit from each SSRC.
-      offer2.sdp = offer2.sdp.replace(/^\r\na=ssrc-group:.*$/gm, '');
-      offer2.sdp = offer2.sdp.replace(/^a=ssrc:[0-9]([0-9]+)(.*)$/gm, 'a=ssrc:$1$2');
+      offer2 = new RTCSessionDescription({
+        type: offer2.type,
+        sdp: offer2.sdp.replace(/^\r\na=ssrc-group:.*$/gm, '')
+          .replace(/^a=ssrc:[0-9]([0-9]+)(.*)$/gm, 'a=ssrc:$1$2')
+      });
     });
 
     // NOTE(mroberts): This is the crux of the issue at the heart of CSDK-1206:
@@ -270,21 +273,24 @@ describe(description, function() {
     //   the right thing to do (especially when we go to Unified Plan SDP) but it's
     //   the way it's worked for a while.
     //
-    (isFirefox || isSafari
+    (isFirefox || isSafari || (isChrome && sdpSemantics === 'unified-plan')
       ? it
       : it.skip
     )('should create a single MediaStreamTrack for each MediaStreamTrack ID in the SDP, regardless of SSRC changes', async () => {
+      const getRemoteTracks = pc => sdpSemantics === 'plan-b'
+        ? flatMap(pc.getRemoteStreams(), stream => stream.getTracks())
+        : flatMap(pc.getTransceivers(), ({ receiver }) => receiver.track);
       const pc = new RTCPeerConnection({ iceServers: [], sdpSemantics });
 
       await pc.setRemoteDescription(offer1);
       const answer1 = await pc.createAnswer();
       await pc.setLocalDescription(answer1);
-      const tracksBefore = flatMap(pc.getRemoteStreams(), stream => stream.getTracks());
+      const tracksBefore = getRemoteTracks(pc);
 
       await pc.setRemoteDescription(offer2);
       const answer2 = await pc.createAnswer();
       await pc.setLocalDescription(answer2);
-      const tracksAfter = flatMap(pc.getRemoteStreams(), stream => stream.getTracks());
+      const tracksAfter = getRemoteTracks(pc);
 
       assert.equal(tracksAfter.length, tracksBefore.length);
       tracksAfter.forEach((trackAfter, i) => {
@@ -642,7 +648,11 @@ function testGetSenders(sdpSemantics, signalingState) {
   context(`"${signalingState}"`, () => {
     it('should return a list of senders', () => {
       const actualSenders = test.peerConnection.getSenders();
-      assert.deepEqual(actualSenders, senders);
+      if (isSafari && signalingState === 'have-local-offer') {
+        assert.equal(actualSenders.length, senders.length + 1);
+      } else {
+        assert.deepEqual(actualSenders, senders);
+      }
     });
   });
 }
@@ -814,7 +824,7 @@ function testDtlsRoleNegotiation(sdpSemantics) {
           });
         });
 
-        (isSafari ? it.skip : it)('RTCPeerConnection 1 answers with "a=setup:passive"', () => {
+        (isSafari && sdpSemantics === 'plan-b' ? it.skip : it)('RTCPeerConnection 1 answers with "a=setup:passive"', () => {
           return pc1.createAnswer().then(answer => {
             assert(answer.sdp.match(/a=setup:passive/));
           });
@@ -860,7 +870,7 @@ function testGlare(sdpSemantics) {
           });
         });
 
-        (isSafari ? it.skip : it)('RTCPeerConnection 1 calls createOffer and setLocalDescription', () => {
+        (isSafari && sdpSemantics === 'plan-b' ? it.skip : it)('RTCPeerConnection 1 calls createOffer and setLocalDescription', () => {
           return pc1.createOffer().then(offer => {
             return pc1.setLocalDescription(offer);
           });